Mark shared files as private, move CLI helpers to shared and create child helper classes for CLI
