+++ vec_inf/shared/_config.py
