+++ vec_inf/shared/_models.py
